Fix handling of ramalama login huggingface
Fixes: https://github.com/containers/ramalama/issues/465
